isolate core options in preparation for categories
