+++ src/mame2003/mame2003.h
@@ -179,9 +179,52 @@ extern void frontend_message_cb(const char *message_string, unsigned frames_to_d
 
 /******************************************************************************
 
-Core options
+  Core options
 
 ******************************************************************************/
+
+enum CORE_OPTIONS  /* controls the order in which core options appear. common, important, and content-specific options should go earlier on the list */
+{
+  OPT_4WAY = 0,
+  OPT_MOUSE_DEVICE,
+  OPT_CROSSHAIR_ENABLED,
+  OPT_SKIP_DISCLAIMER,
+  OPT_SKIP_WARNINGS,
+  OPT_DISPLAY_SETUP,
+  OPT_NEOGEO_BIOS,
+  OPT_STV_BIOS,
+  OPT_USE_ALT_SOUND,
+  OPT_SHARE_DIAL,
+  OPT_DEADZONE,
+  OPT_VECTOR_RESOLUTION,
+  OPT_VECTOR_ANTIALIAS,
+  OPT_VECTOR_BEAM,
+  OPT_VECTOR_TRANSLUCENCY,
+  OPT_VECTOR_FLICKER,
+  OPT_VECTOR_INTENSITY,
+  OPT_CORE_SYS_SUBFOLDER,
+  OPT_CORE_SAVE_SUBFOLDER,
+  OPT_AUTOSAVE_HISCORE,
+  OPT_TATE_MODE,
+  OPT_BRIGHTNESS,
+  OPT_GAMMA,
+  OPT_FRAMESKIP,
+  OPT_CPU_CLOCK_SCALE,
+  OPT_SAMPLE_RATE,
+  OPT_INPUT_INTERFACE,
+  OPT_MAME_REMAPPING,
+  OPT_ARTWORK,
+  OPT_ART_RESOLUTION,
+  OPT_ART_OVERLAY_OPACITY,
+  OPT_NVRAM_BOOTSTRAP,
+  OPT_CHEAT_INPUT_PORTS,
+  OPT_MACHINE_TIMING,
+  OPT_DIGITAL_JOY_CENTERING,
+#if (HAS_CYCLONE || HAS_DRZ80)
+  OPT_CYCLONE_MODE,
+#endif
+  OPT_end /* dummy last entry */
+};
 
 
 /******************************************************************************
